FP-5 updated for a mysql db in place of h2

+++ src/main/java/com/finca_la_caprichosa/model/Goat.java
@@ -2,6 +2,8 @@ package com.finca_la_caprichosa.model;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.NamedQueries;
 import javax.persistence.NamedQuery;
@@ -21,7 +23,7 @@ import java.io.Serializable;
 public class Goat implements Serializable {
 
     @Id
-    private Integer id;
+    private Long id;
 
     @NotNull
     @Column(updatable = false)
@@ -30,11 +32,11 @@ public class Goat implements Serializable {
     @NotNull
     private boolean producing;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
